Basic Datapage Styling
Unfortunately, getting ElementRef, ViewChild, and/or Renderer2 to set the height of the iframe was a bust. I couldn't even get a basic databind of the src URLs working without DOM sanitization. The scope of the variables never cooperated outside of the ngOnInit function and I couldn't figure out how to pull those operations into the button click functions. Here are some leftover notes from my adventures to give you an idea of what I was trying to do.

- use renderer2 to generate the whole iframe element including styling before putting it into the HTML file
- Cross-domain iframes may require a postMessage option

  const framer = this.renderer.createElement('iframe');
  this.renderer.setAttribute(framer, 'src', frameSource);
  this.renderer.setStyle(framer, 'height', framer.contentWindow.document.documentElement.scrollHeight);